Runtime Crash in SkillDataManager.java traced back to Animations.java. Found that there was a problem with Steel Whirlwind animation(crash at line 1648 when calling getChargingPower)
Problem: When instanceof is used a nullpointer was created, and subsequently a null PlayerPatch was passed and then an uninitiated skillcontainer was used. Example: if(entitypatch instance of PlayerPatch<?> playerpatch): If playerpatch is referenced, it would be a nullpointer.

Solution: Remove the object name from if statement (e.g. playerpatch) and in the first line under if statement typecast the entityPatch to a PlayerPatch.

Similar issue found in other if statement, therefore all others were adjusted. Some may not have needed the audjustment, but as I am not familiar enough with the codebase and not too knowledgeable, I did it anyways, as it made sense to me to do so.