Modified line 536 of MobPatchReloadListener.java to take entry.getKey… (#1665)
* Modified line 536 of MobPatchReloadListener.java to take entry.getKey() as a parameter instead of the entry itself since entry is a K,V hashmap node instead of a EntityType<?>.

I assume this will resolve issue of not being able to join multiplayer servers.

* Runtime Crash in SkillDataManager.java traced back to Animations.java. Found that there was a problem with Steel Whirlwind animation(crash at line 1648 when calling getChargingPower)

Problem: When instanceof is used a nullpointer was created, and subsequently a null PlayerPatch was passed and then an uninitiated skillcontainer was used. Example: if(entitypatch instance of PlayerPatch<?> playerpatch): If playerpatch is referenced, it would be a nullpointer.

Solution: Remove the object name from if statement (e.g. playerpatch) and in the first line under if statement typecast the entityPatch to a PlayerPatch.

Similar issue found in other if statement, therefore all others were adjusted. Some may not have needed the audjustment, but as I am not familiar enough with the codebase and not too knowledgeable, I did it anyways, as it made sense to me to do so.

* Revert "Runtime Crash in SkillDataManager.java traced back to Animations.java. Found that there was a problem with Steel Whirlwind animation(crash at line 1648 when calling getChargingPower)"

This reverts commit c29adcfdc3cc983e2eeff5f88fdab4a464d4b554.